backend/OffchainPriceProvider/requirements.txt
# Web server
Flask==2.2.3
Flask-CORS==3.0.10
gunicorn==20.1.0

# Data manipulation
pandas==1.5.3

# HTTP requests
requests==2.28.2

# Ethereum
web3==6.0.0

# Environment variables
python-dotenv==1.0.0
